project completed

+++ backend/controllers/product.controller.js
@@ -1,10 +1,10 @@
-import express from "express";
-import mongoose from 'mongoose';
-import Product from '../models/product.model';
+import mongoose from "mongoose";
+import Product from "../models/product.model.js";
 
-const router = express.Router();
 
-router.get("/", async (req, res) => {
+
+
+export const getProducts = async (req, res) => {
     try {
         const products = await Product.find({});
         res.status(200).json({ success: true, data: products });
@@ -13,10 +13,11 @@ router.get("/", async (req, res) => {
         res.status(500).json({ success: false, message: "Server Error"});
         
     }
-});
+};
 
 
-router.post("/", async (req,res) => {
+
+export const createProduct = async (req,res) => {
     const product = req.body; // user will send this data
 
     if(!product.name || !product.price || !product.image) {
@@ -32,10 +33,10 @@ router.post("/", async (req,res) => {
         console.error("Error in Create Product:", error.message);
         res.status(500).json({success: false, message: "Server Error"});
     }
-});
+};
 
 
-router.put("/:id", async (req,res) => {
+export const updateProduct = async (req,res) => {
     const { id } = req.params;
     const product = req.body;
 
@@ -72,18 +73,25 @@ router.put("/:id", async (req,res) => {
             message: "Server Error"
         });
     }
-});
+};
 
-router.delete("/:id", async (req, res) => {
+
+
+export const deleteProduct = async (req, res) => {
     const { id } = req.params;
+
+    if(!mongoose.Types.ObjectId.isValid(id)){
+        return res.status(404).json({ 
+            success: false, 
+            message: "Invalid Product Id"
+        });
+    }
 
     try {
         await Product.findByIdAndDelete(id);
         res.status(200).json({ success: true, message: "Product Deleted"});
     } catch (error) {
-        
-        res.status(404).json({success: false, message: "Product not found"});
+        console.log("error in deleting product:", error.message);
+        res.status(500).json({success: false, message: "Server Error"});
     }
-});
-
-export default router;
+};
